+++ crates/daemon-server/tests/watcher_test.rs
@@ -21,11 +21,7 @@ struct RecordingWatcher;
 
 #[async_trait]
 impl FileWatcher<TestState> for RecordingWatcher {
-    async fn on_file_event(
-        &mut self,
-        state: &mut TestState,
-        event: &FileEvent,
-    ) -> miette::Result<()> {
+    async fn on_file_event(&mut self, state: TestState, event: &FileEvent) -> miette::Result<()> {
         state
             .events
             .lock()
@@ -40,11 +36,7 @@ struct FailingWatcher;
 
 #[async_trait]
 impl FileWatcher<TestState> for FailingWatcher {
-    async fn on_file_event(
-        &mut self,
-        _state: &mut TestState,
-        _event: &FileEvent,
-    ) -> miette::Result<()> {
+    async fn on_file_event(&mut self, _state: TestState, _event: &FileEvent) -> miette::Result<()> {
         Err(miette::miette!("deliberate failure"))
     }
 }
@@ -56,11 +48,7 @@ struct SecondRecordingWatcher {
 
 #[async_trait]
 impl FileWatcher<TestState> for SecondRecordingWatcher {
-    async fn on_file_event(
-        &mut self,
-        state: &mut TestState,
-        event: &FileEvent,
-    ) -> miette::Result<()> {
+    async fn on_file_event(&mut self, state: TestState, event: &FileEvent) -> miette::Result<()> {
         state
             .events
             .lock()