+++ crates/task-runner/src/lib.rs
@@ -4,7 +4,6 @@ pub mod command_executor;
 pub mod output_archiver;
 pub mod output_hydrater;
 mod run_state;
-mod task_hashing;
 mod task_runner;
 mod task_runner_error;