+++ crates/task-runner/src/output_archiver.rs
@@ -1,21 +1,22 @@
 use crate::task_runner_error::TaskRunnerError;
+use miette::IntoDiagnostic;
 use moon_app_context::AppContext;
 use moon_common::color;
-use moon_project::Project;
 use moon_remote::{ActionState, RemoteService};
 use moon_task::Task;
 use starbase_archive::Archiver;
 use starbase_archive::tar::TarPacker;
 use std::path::{Path, PathBuf};
+use std::sync::Arc;
+use tokio::task::spawn_blocking;
 use tracing::{debug, instrument, warn};
 
 /// Cache outputs to the `.moon/cache/outputs` folder and to the cloud,
 /// so that subsequent builds are faster, and any local outputs
 /// can be hydrated easily.
 pub struct OutputArchiver<'task> {
-    pub app_context: &'task AppContext,
-    pub project: &'task Project,
-    pub task: &'task Task,
+    pub app_context: &'task Arc<AppContext>,
+    pub task: &'task Arc<Task>,
 }
 
 impl OutputArchiver<'_> {
@@ -45,8 +46,7 @@ impl OutputArchiver<'_> {
                 hash, "Archiving task outputs from project"
             );
 
-            self.create_local_archive(hash, &archive_file)?;
-            archived = true;
+            archived = self.create_local_archive(hash, &archive_file).await?;
         } else {
             debug!(
                 task_target = self.task.target.as_str(),
@@ -109,47 +109,57 @@ impl OutputArchiver<'_> {
     }
 
     #[instrument(skip(self))]
-    fn create_local_archive(&self, hash: &str, archive_file: &Path) -> miette::Result<()> {
+    async fn create_local_archive(&self, hash: &str, archive_file: &Path) -> miette::Result<bool> {
         debug!(
             task_target = self.task.target.as_str(),
             hash,
             archive_file = ?archive_file, "Creating archive file"
         );
 
-        // Create the archiver instance based on task outputs
-        let mut archive = Archiver::new(&self.app_context.workspace_root, archive_file);
+        // Clone values to run in a blocking thread
+        let archive_file = archive_file.to_path_buf();
+        let app_context = Arc::clone(self.app_context);
+        let task = Arc::clone(self.task);
+        let hash = hash.to_string();
 
-        for output_file in self.task.output_files.keys() {
-            archive.add_source_file(output_file.as_str(), None);
-        }
+        let archived = spawn_blocking(move || {
+            // Create the archiver instance based on task outputs
+            let mut archive = Archiver::new(&app_context.workspace_root, &archive_file);
 
-        for output_glob in self.task.output_globs.keys() {
-            archive.add_source_glob(output_glob.as_str());
-        }
+            for output_file in task.output_files.keys() {
+                archive.add_source_file(output_file.as_str(), None);
+            }
 
-        // Also include stdout/stderr logs in the tarball
-        let state_dir = self
-            .app_context
-            .cache_engine
-            .state
-            .get_target_dir(&self.task.target);
+            for output_glob in task.output_globs.keys() {
+                archive.add_source_glob(output_glob.as_str());
+            }
 
-        archive.add_source_file(state_dir.join("stdout.log"), None);
+            // Also include stdout/stderr logs in the tarball
+            let state_dir = app_context.cache_engine.state.get_target_dir(&task.target);
 
-        archive.add_source_file(state_dir.join("stderr.log"), None);
+            archive.add_source_file(state_dir.join("stdout.log"), None);
 
-        // Pack the archive
-        if let Err(error) = archive.pack(TarPacker::new_gz) {
-            warn!(
-                task_target = self.task.target.as_str(),
-                hash,
-                archive_file = ?archive_file,
-                "Failed to package outputs into archive: {}",
-                color::muted_light(error.to_string()),
-            );
-        }
+            archive.add_source_file(state_dir.join("stderr.log"), None);
 
-        Ok(())
+            // Pack the archive
+            if let Err(error) = archive.pack(TarPacker::new_gz) {
+                warn!(
+                    task_target = task.target.as_str(),
+                    hash,
+                    archive_file = ?archive_file,
+                    "Failed to package outputs into archive: {}",
+                    color::muted_light(error.to_string()),
+                );
+
+                return false;
+            }
+
+            true
+        })
+        .await
+        .into_diagnostic()?;
+
+        Ok(archived)
     }
 
     #[instrument(skip(self, state))]